# Conflicts:
#   BE/LOFI/src/main/java/com/ssafy/lofi/dto/response/MissingAnimalAPIResponse.java
#   BE/LOFI/src/main/resources/application.properties

+++ BE/LOFI/src/main/java/com/ssafy/lofi/db/repository/MissingAnlmalRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 
-public interface MissingAnlmalRepository extends JpaRepository<MissingAnimal, Integer> {
+public interface MissingAnlmalRepository extends JpaRepository<MissingAnimal, Long> {
     @Modifying
     @Query(value = "truncate table missing_animal", nativeQuery = true)
     void truncateMissingPerson();
